requirements.txt
# Kafka
kafka-python==2.2.15

# API e requisições
requests==2.32.5

# Banco de dados
sqlalchemy==2.0.44

# Dashboard
streamlit==1.50.0
plotly==6.3.1.
pandas==2.3.3

# Variáveis de ambiente
python-dotenv==1.1.1